+++ test/1-main.test.js
@@ -1,3 +1,2 @@
-global.clone = require('../src/clone.js');
-
+global.clone   = require('../src/clone.js');
 module.exports = require('./clone.spec.js');